requirements.txt
gradio>=4.0.0
requests>=2.31.0
pandas>=2.0.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langchain-groq>=0.0.1
langgraph>=0.0.20
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
supabase>=2.0.0
tavily-python>=0.3.0
wikipedia>=1.4.0
arxiv>=2.1.0